+++ src/reducers/search.js
@@ -9,12 +9,39 @@ import {
   OPEN_SEARCHDRAWER,
   CLOSE_SEARCHDRAWER,
 } from '../actions/search.js'
+import { CF_RECEIVE_PAGE } from '../actions/contentful/page'
 
 const localSearchPref = localStorage.getItem('searchPreference')
+
+// Sets search type based on a priority sequence. Note: This modifies the
+// newState passed in, so assumes you've already created a copy of the state
+const setSearchType = (newState) => {
+  // User's session
+  if (newState.sessionPref) {
+    newState.searchType = newState.sessionPref
+    return
+  }
+
+  // User's saved preference
+  if (newState.hasPref && newState.usePref) {
+    newState.searchType = newState.pref.uid
+    return
+  }
+
+  // Preference defined by the page/content
+  if (newState.pageSearchPref) {
+    newState.searchType = newState.pageSearchPref
+    return
+  }
+
+  // Fallback default
+  newState.searchType = 'ONESEARCH'
+}
+
 export default (
   state = {
     drawerOpen: true,
-    searchType: localSearchPref || 'ONESEARCH',
+    searchType: 'ONESEARCH',
     searchBoxOpen: false,
     advancedSearch: false,
     hasPref: localSearchPref !== null,
@@ -25,10 +52,13 @@ export default (
 ) => {
   switch (action.type) {
     case SET_SEARCH:
-      return Object.assign({}, state, {
-        searchType: action.searchType,
-        usePref: false,
-      })
+      {
+        let newState = Object.assign({}, state, {
+          sessionPref: action.searchType,
+        })
+        setSearchType(newState)
+        return newState
+      }
     case OPEN_SEARCHBOX:
       return Object.assign({}, state, { searchBoxOpen: true })
     case CLOSE_SEARCHBOX:
@@ -60,6 +90,14 @@ export default (
       return Object.assign({}, state, { drawerOpen: true })
     case CLOSE_SEARCHDRAWER:
       return Object.assign({}, state, { drawerOpen: false })
+    case CF_RECEIVE_PAGE:
+      {
+        let newState = Object.assign({}, state, {
+          pageSearchPref: action.page.fields.defaultSearchScope,
+        })
+        setSearchType(newState)
+        return newState
+      }
     default:
       return state
   }